requirements.txt
annotated-types==0.6.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
ChatterBot2==1.1.0a7
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.4.1
confection==0.1.4
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
mathparse==0.1.2
murmurhash==1.0.10
numpy==1.24.4
packaging==24.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.0
pydantic_core==2.18.1
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2024.1
pywin32==306
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.3
SQLAlchemy==1.3.24
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
